changed return on divide method

+++ src/main/java/com/example/springdocker/model/MyMathCalc.java
@@ -10,7 +10,10 @@ public class MyMathCalc {
         return a * b;
     }
 
-    public float divide(float a, float b) {
-        return a/b;
+
+    public double divide(double a, double b) {
+        return a/b ;
+
+ 
     }
 }
